# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type AudioProjectStatus = typeof AudioProjectStatus[keyof typeof AudioProjectStatus];


export const AudioProjectStatus = {
  ready: 'ready',
  processing: 'processing',
  done: 'done',
  error: 'error',
} as const;

export interface EnhancementSettings {
  /** @nullable */
  eqPresetId: string | null;
  /** 0.0 to 2.0, 1.0 = original */
  stereoWidth: number;
  /** 0.0 to 1.0 */
  clarityAmount: number;
  humReduction: boolean;
  /**
     * 50 or 60 Hz
     * @nullable
     */
  humFrequency?: number | null;
  /** 0.0 to 1.0 */
  noiseReduction: number;
  /** 0.0 to 1.0 */
  sibilanceReduction: number;
  clipRepair: boolean;
  preRingFix: boolean;
  enabled: boolean;
}

export interface MasteringSettings {
  /** @nullable */
  genreId: string | null;
  /** 0.0 to 1.0 */
  compressionAmount: number;
  /** 0.0 to 1.0 */
  dynamicEqAmount: number;
  /** 0.0 to 1.0 */
  exciterAmount: number;
  /** Target integrated loudness, e.g. -14 for Spotify */
  targetLufs: number;
  enabled: boolean;
}

export interface AudioProject {
  id: string;
  name: string;
  originalFilename: string;
  fileUrl: string;
  status: AudioProjectStatus;
  createdAt: string;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  sampleRate?: number | null;
  enhancementSettings: EnhancementSettings;
  masteringSettings: MasteringSettings;
}

export interface AudioProjectRename {
  /** @minLength 1 */
  name: string;
}

export interface EnhancementSettingsInput {
  /** @nullable */
  eqPresetId?: string | null;
  stereoWidth?: number;
  clarityAmount?: number;
  humReduction?: boolean;
  /** @nullable */
  humFrequency?: number | null;
  noiseReduction?: number;
  sibilanceReduction?: number;
  clipRepair?: boolean;
  preRingFix?: boolean;
  enabled?: boolean;
}

export interface MasteringSettingsInput {
  /** @nullable */
  genreId?: string | null;
  compressionAmount?: number;
  dynamicEqAmount?: number;
  exciterAmount?: number;
  targetLufs?: number;
  enabled?: boolean;
}

export type EqBandType = typeof EqBandType[keyof typeof EqBandType];


export const EqBandType = {
  lowshelf: 'lowshelf',
  highshelf: 'highshelf',
  peaking: 'peaking',
  lowpass: 'lowpass',
  highpass: 'highpass',
  notch: 'notch',
} as const;

export interface EqBand {
  frequency: number;
  gain: number;
  q: number;
  type: EqBandType;
}

export type EqPresetCategory = typeof EqPresetCategory[keyof typeof EqPresetCategory];


export const EqPresetCategory = {
  voice: 'voice',
  music: 'music',
  podcast: 'podcast',
  broadcast: 'broadcast',
  film: 'film',
  restoration: 'restoration',
} as const;

export interface EqPreset {
  id: string;
  name: string;
  description: string;
  category: EqPresetCategory;
  bands: EqBand[];
}

export interface MasteringGenre {
  id: string;
  name: string;
  description: string;
  targetLufs: number;
  compressionAmount: number;
  dynamicEqAmount: number;
  exciterAmount: number;
  character?: string;
}