# lib/api-zod/src/generated/types/audioProject.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { AudioProjectStatus } from './audioProjectStatus';
import type { EnhancementSettings } from './enhancementSettings';
import type { MasteringSettings } from './masteringSettings';

export interface AudioProject {
  id: string;
  name: string;
  originalFilename: string;
  fileUrl: string;
  status: AudioProjectStatus;
  createdAt: Date;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  sampleRate?: number | null;
  enhancementSettings: EnhancementSettings;
  masteringSettings: MasteringSettings;
}