# lib/api-zod/src/generated/types/enhancementSettings.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface EnhancementSettings {
  /** @nullable */
  eqPresetId: string | null;
  /** 0.0 to 2.0, 1.0 = original */
  stereoWidth: number;
  /** 0.0 to 1.0 */
  clarityAmount: number;
  humReduction: boolean;
  /**
     * 50 or 60 Hz
     * @nullable
     */
  humFrequency?: number | null;
  /** 0.0 to 1.0 */
  noiseReduction: number;
  /** 0.0 to 1.0 */
  sibilanceReduction: number;
  clipRepair: boolean;
  preRingFix: boolean;
  enabled: boolean;
}