# lib/api-zod/src/generated/types/enhancementSettingsInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface EnhancementSettingsInput {
  /** @nullable */
  eqPresetId?: string | null;
  stereoWidth?: number;
  clarityAmount?: number;
  humReduction?: boolean;
  /** @nullable */
  humFrequency?: number | null;
  noiseReduction?: number;
  sibilanceReduction?: number;
  clipRepair?: boolean;
  preRingFix?: boolean;
  enabled?: boolean;
}