# lib/api-zod/src/generated/types/eqBand.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { EqBandType } from './eqBandType';

export interface EqBand {
  frequency: number;
  gain: number;
  q: number;
  type: EqBandType;
}