# lib/api-zod/src/generated/types/eqPreset.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { EqBand } from './eqBand';
import type { EqPresetCategory } from './eqPresetCategory';

export interface EqPreset {
  id: string;
  name: string;
  description: string;
  category: EqPresetCategory;
  bands: EqBand[];
}