# lib/api-zod/src/generated/types/masteringGenre.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface MasteringGenre {
  id: string;
  name: string;
  description: string;
  targetLufs: number;
  compressionAmount: number;
  dynamicEqAmount: number;
  exciterAmount: number;
  character?: string;
}