# lib/api-zod/src/generated/types/masteringSettings.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface MasteringSettings {
  /** @nullable */
  genreId: string | null;
  /** 0.0 to 1.0 */
  compressionAmount: number;
  /** 0.0 to 1.0 */
  dynamicEqAmount: number;
  /** 0.0 to 1.0 */
  exciterAmount: number;
  /** Target integrated loudness, e.g. -14 for Spotify */
  targetLufs: number;
  enabled: boolean;
}