# lib/api-zod/src/generated/types/masteringSettingsInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface MasteringSettingsInput {
  /** @nullable */
  genreId?: string | null;
  compressionAmount?: number;
  dynamicEqAmount?: number;
  exciterAmount?: number;
  targetLufs?: number;
  enabled?: boolean;
}